第七模块-项目实战CRM-爬虫开发/爬虫&APP逆向/day01 x视频（课件和代码）/代码/day01/requirements.txt
pycryptodome==3.10.1
PyExecJS==1.5.1
requests==2.18.0
